# Microoptimize the walk() function