@kmod 2025-01-08
Haha yeah at first I got worried about how I broke so much without realizing it

Also wanted to say, thanks for the project! It's been very handy